componentize brainstorming
